Add external types (#109)
This PR adds external types to Orng. They are declared the same way as
normal external types. If an initializer isn't given, they are
completely opaque, but still useable. If an initializer is given, the
compiler will allow the user to work with them as if they were defined
within Orng, but output them as C external types.