Add incremental builds (#123)
This PR makes it so that the compiler will only re-generate and
re-compile the C files when a module's hash has changed or one of its
dependencies has changed, and will only re-link packages when a package
has changed or one of its dependencies has changed. Also added an `orng
clean`, which clears the cache for the package.